register component

+++ backend/src/main/java/com/pieces/album/model/Photo.java
@@ -16,7 +16,7 @@ public class Photo {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int photoId;
+    private Long photoId;
 
     private String photoName;
 
@@ -39,11 +39,11 @@ public class Photo {
     private List<Comment> commentList;
 
 
-    public int getPhotoId() {
+    public Long getPhotoId() {
         return photoId;
     }
 
-    public void setPhotoId(int photoId) {
+    public void setPhotoId(Long photoId) {
         this.photoId = photoId;
     }
